fix: update SQLite migrations to generate same Go types as MySQL/PostgreSQL
Changes made to ensure GORM generates identical struct types across all database backends:
- Changed INTEGER to BIGINT for all timestamp fields (create_time_since_epoch, last_update_time_since_epoch, milliseconds_since_epoch)
- Changed INTEGER to BOOLEAN for all boolean fields (is_custom_property, bool_value, is_index_step)
- Restored SQLite schema validation in CI workflow

This ensures SQLite generates the same Go types:
- BIGINT maps to int64 (matching MySQL bigint)
- BOOLEAN maps to bool (matching MySQL tinyint(1))
- Maintains schema consistency across all database backends

+++ internal/datastore/embedmd/sqlite/migrations/000001_create_artifact_table.up.sql
@@ -6,8 +6,8 @@ CREATE TABLE IF NOT EXISTS "Artifact" (
     state INTEGER DEFAULT NULL,
     name TEXT DEFAULT NULL,
     external_id TEXT DEFAULT NULL,
-    create_time_since_epoch INTEGER NOT NULL DEFAULT '0',
-    last_update_time_since_epoch INTEGER NOT NULL DEFAULT '0',
+    create_time_since_epoch BIGINT NOT NULL DEFAULT '0',
+    last_update_time_since_epoch BIGINT NOT NULL DEFAULT '0',
     UNIQUE (external_id),
     UNIQUE (type_id, name)
 );
